refactor(plugin): move enableNativeTests out of the public plugin API
The Cordierite XCTest wiring was only ever meant for our own testing,
not for consumer apps. Replace the enableNativeTests plugin option
with a playground-local config plugin that injects the same Podfile
test-spec line, so the public @cordierite/react-native plugin no
longer exposes it.
